// API/src/database/types/job-repository.d.ts
export type jobRepository = {
    id: number;
    company_name: string;
    position: string;
    seniority_level: string;
    payment_currency: string;
    initial_salary: number;
    current_salary: number;
    vacancy_modality: string;
    work_regime: string;
    place: string;
    status: string;
    created_at: string;
    updated_at: string;
};